backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
python-jose[cryptography]>=3.3.0
ptyprocess>=0.7.0
pyyaml>=6.0
apscheduler>=3.10.0
